supplementary/figS34: add material
This computation shall check whether the force amplitude for the VLP source as
inferred from teleseismic recordings would be consistent with the simulated
seiche amplitude. The computation is discussed in section 'vii) The seiche as
a potential source of the VLP signal' of the supplementary material.

One of the sketches is used in Fig. S34 in the supplementary material.